feat: modify Navbar to include getSessionUser() API
