+++ src/components/Navbar/server.tsx
@@ -1,0 +1,7 @@
+import { NavbarItems } from "./client";
+import { getSessionUser } from "@/lib/tools/session/api";
+
+export const LoadingNav = async () => {
+  const user = await getSessionUser();
+  return <NavbarItems loggedIn={!!user} />;
+};
